+++ aerogear-android-push-test/src/main/java/org/jboss/aerogear/android/unifiedpush/test/util/StubActivity.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.aerogear.android.impl.util;
+package org.jboss.aerogear.android.authentication.test.util;
 
 import android.app.Activity;
 import android.os.Bundle;
