+++ aerogear-android-push/src/main/java/org/jboss/aerogear/android/unifiedpush/RegistrarManager.java
@@ -27,7 +27,7 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import org.jboss.aerogear.android.ConfigurationProvider;
+import org.jboss.aerogear.android.core.ConfigurationProvider;
 
 /**
  * This is the factory and accessors for PushRegistrars
